tests/wpt/tests/html/canvas/element/manual/draw-element-image/privacy/svg-mask-filter.tentative.https.sub.html

<!DOCTYPE html>
<html>
<head>
  <title>drawElementImage does not use cross-origin feImages on filtered SVG text inside masks</title>
  <link rel="help" href="https://github.com/WICG/html-in-canvas">
  <script src='/resources/testharness.js'></script>
  <script src='/resources/testharnessreport.js'></script>
  <script src='/html/canvas/resources/canvas-tests.js'></script>
  <script src="/html/canvas/resources/wait-for-canvas-paint.js"></script>
</head>

<body>
  <svg width="0" height="0">
    <defs>
      <filter id="filter-same" filterUnits="userSpaceOnUse" x="0" y="0" width="100" height="100">
        <feImage href="https://{{location[host]}}/images/green-100x100.png"/>
        <feComponentTransfer>
          <feFuncR type="linear" slope="0" intercept="1"/>
          <feFuncG type="linear" slope="0" intercept="1"/>
          <feFuncB type="linear" slope="0" intercept="1"/>
        </feComponentTransfer>
      </filter>
      <filter id="filter-cross" filterUnits="userSpaceOnUse" x="0" y="0" width="100" height="100">
        <feImage href="https://{{hosts[alt][www]}}:{{ports[https][0]}}/images/red-100x100.png"/>
        <feComponentTransfer>
          <feFuncR type="linear" slope="0" intercept="1"/>
          <feFuncG type="linear" slope="0" intercept="1"/>
          <feFuncB type="linear" slope="0" intercept="1"/>
        </feComponentTransfer>
      </filter>
      <mask id="mask-same" maskUnits="userSpaceOnUse" x="0" y="0" width="100" height="100">
        <text x="0" y="50" filter="url(#filter-same)">x</text>
      </mask>
      <mask id="mask-cross" maskUnits="userSpaceOnUse" x="0" y="0" width="100" height="100">
        <text x="0" y="50" filter="url(#filter-cross)">x</text>
      </mask>
    </defs>
  </svg>

  <canvas id="canvas" width="100" height="200" layoutsubtree>
    <svg id="child" width="100" height="200">
      <svg x="0" y="0" width="100" height="100">
        <rect width="100" height="100" fill="rgb(30, 130, 230)" mask="url(#mask-same)"/>
      </svg>
      <svg x="0" y="100" width="100" height="100">
        <rect width="100" height="100" fill="rgb(40, 140, 240)" mask="url(#mask-cross)"/>
      </svg>
    </svg>
  </canvas>

  <script>
    window.onload = () => {
      promise_test(async function(t) {
        await waitForCanvasPaint(canvas);
        var ctx = canvas.getContext('2d');
        ctx.fillStyle = 'blue';
        ctx.fillRect(0, 0, 100, 200);
        ctx.drawElementImage(document.getElementById('child'), 0, 0);

        // Fetch all pixel data once to avoid multiple slow readbacks.
        const imgData = ctx.getImageData(0, 0, canvas.width, canvas.height);

        let pixel = _getPixelFromImageData(imgData, 50, 50);
        assert_array_equals(pixel, [30, 130, 230, 255], "Same origin filter image in external mask should be drawn");

        pixel = _getPixelFromImageData(imgData, 50, 150);
        assert_array_equals(pixel, [0, 0, 255, 255], "Cross origin filter image in external mask should not be drawn");
      });
    }
  </script>
</body>
</html>
